backend/requirements.txt
flask
pandas
plotly
psycopg2-binary
pytest
python-dotenv
requests
simplejson
streamlit
unidecode